Moved some files around, visitor imps now in own file
