+++ src/final/CMakeLists.txt
@@ -3,7 +3,7 @@ project(final)
 add_executable(${PROJECT_NAME} main.cpp initialization.cpp initialization.hpp main_network_worker.cpp main_network_worker.hpp
         ../server/network_worker.cpp ../server/network_worker.hpp ../server/worker.hpp pins.hpp main_worker.cpp main_worker.hpp
 	main_buff_logger.cpp main_buff_logger.hpp timed_item.hpp timed_item.cpp ini_config.cpp ini_config.hpp timed_item_list.cpp
-	timed_item_list.hpp ../server/queue_visitor.hpp ../server/queue_visitor.cpp)
+	timed_item_list.hpp ../server/queue_visitor.hpp ../server/queue_visitor_imps.cpp)
 
 SET(FINAL_LIBS_NO_FAKED timestamps.a listener.a circ_buff.a logger.a)
 SET(FINAL_LIBS ${FINAL_LIBS_NO_FAKED} adc_lib.a)
@@ -25,7 +25,7 @@ find_package(Boost COMPONENTS program_options REQUIRED)
 include_directories(${Boost_INCLUDE_DIRS})
 add_executable(final_mocked main.cpp ../../test/mocks/fake_initialization.cpp main_network_worker.cpp main_network_worker.hpp
         ../server/network_worker.cpp main_worker.cpp main_buff_logger.cpp timed_item.cpp ini_config.cpp timed_item_list.cpp timed_item_list.hpp
-		../server/queue_visitor.hpp ../server/queue_visitor.cpp)
+		../server/queue_visitor.hpp ../server/queue_visitor_imps.hpp ../server/queue_visitor_imps.cpp)
 
 #add_dependencies(final_mocked ${Boost_LIBRARIES} ${FINAL_LIBS_MOCKS})
 target_link_libraries(final_mocked ${Boost_LIBRARIES} ${FINAL_LIBS_MOCKS} pthread)
