+++ src/server/queue_visitor.hpp
@@ -1,3 +1,8 @@
+/*
+ * Defines interfaces for work_queue visitors and network_queue visitors
+ * (latter still TODO).
+ */
+
 #ifndef SOFTWARE_QUEUE_VISITOR_HPP
 #define SOFTWARE_QUEUE_VISITOR_HPP
 
@@ -83,51 +88,6 @@ class work_queue_visitor {
                 }
             }
         }
-};
-
-/**
- * Work queue visitor that performs the standard tasks for each action.
- */
-class main_work_queue_visitor : public work_queue_visitor {
-    public:
-        void visitProc(work_queue_item&);
-
-        virtual void visitTimed(work_queue_item&); // virtual because overridden by titan_work_queue_visitor
-
-        void visitIgn(work_queue_item&);
-
-        void visitNone(work_queue_item&);
-
-        void visitDefault(work_queue_item&);
-
-        main_work_queue_visitor(safe_queue<work_queue_item>& qw,
-                                safe_queue<network_queue_item>& qn,
-                                adc_block& adcs,
-                                main_network_worker* nw_ref)
-                                : work_queue_visitor(qw, qn, adcs, nw_ref)
-        {
-
-        }
-        ~main_work_queue_visitor() { }
-};
-
-/**
- * Work queue visitor for Titan. Inherits most of its code from
- * main_work_queue_visitor, just with different ignition procedures.
- */
-class titan_work_queue_visitor : public main_work_queue_visitor {
-    public:
-        void visitTimed(work_queue_item&);
-
-        titan_work_queue_visitor(safe_queue<work_queue_item>& qw,
-                                 safe_queue<network_queue_item>& qn,
-                                 adc_block& adcs,
-                                 main_network_worker* nw_ref)
-                                 : main_work_queue_visitor(qw, qn, adcs, nw_ref)
-        {
-
-        }
-        ~titan_work_queue_visitor() { }
 };
 
 /**
